+++ app/concepts/stock/operation/load_etablissement.rb
@@ -1,0 +1,13 @@
+class Stock
+  module Operation
+    class LoadEtablissement < Trailblazer::Operation
+      step :stock_model
+      step Nested Task::FetchLatestRemoteStockEtablissement
+      step Nested Task::Load
+
+      def stock_model(ctx, **)
+        ctx[:stock_model] = StockEtablissement
+      end
+    end
+  end
+end
