--- app/concepts/unite_legale/operation/load.rb
@@ -1,13 +1,0 @@
-class UniteLegale
-  module Operation
-    class Load < Trailblazer::Operation
-      step :stock_model
-      step Nested Task::FetchLatestRemoteStock
-      step Nested Stock::Task::Load
-
-      def stock_model(ctx, **)
-        ctx[:stock_model] = StockUniteLegale
-      end
-    end
-  end
-end
