+++ spec/concepts/stock/operation/update_database_spec.rb
@@ -8,12 +8,12 @@ describe Stock::Operation::UpdateDatabase, :trb, vcr: { cassette_name: 'update_d
   it { is_expected.to be_success }
 
   it 'calls LoadUniteLegale' do
-    expect_to_call_nested_operation(UniteLegale::Operation::Load)
+    expect_to_call_nested_operation(Stock::Operation::LoadEtablissement)
     subject
   end
 
   it 'calls LoadEtablissements' do
-    expect_to_call_nested_operation(Etablissement::Operation::Load)
+    expect_to_call_nested_operation(Stock::Operation::LoadUniteLegale)
     subject
   end
 end
