+++ spec/concepts/stock/task/fetch_latest_remote_stock_etablissement_spec.rb
@@ -1,28 +1,28 @@
 require 'rails_helper'
 
-describe UniteLegale::Task::FetchLatestRemoteStock do
+describe Stock::Task::FetchLatestRemoteStockEtablissement do
   subject { described_class.call logger: logger }
 
   let(:logger) { instance_spy Logger }
 
-  describe 'valid HTTP interaction', vcr: { cassette_name: 'data_gouv_sirene_july_OK' } do
-    let(:latest_remote_stock_file) { 'http://files.data.gouv.fr/insee-sirene/StockUniteLegale_utf8.zip' }
+  describe 'valid HTTP interaction', vcr: { cassette_name: 'cquest_geo_sirene_may_OK' } do
+    let(:latest_remote_stock_file) { 'http://data.cquest.org/geo_sirene/v2019/2019-05/StockEtablissement_utf8_geo.csv.gz' }
 
     it { is_expected.to be_success }
 
-    its([:remote_stock]) { is_expected.to be_a StockUniteLegale }
+    its([:remote_stock]) { is_expected.to be_a StockEtablissement }
 
     it 'returns a stock form the uri' do
       expect(subject[:remote_stock]).to have_attributes(
         year: '2019',
-        month: '07',
+        month: '05',
         status: 'PENDING',
         uri: latest_remote_stock_file
       )
     end
   end
 
-  context 'when HTTP interaction fails', vcr: { cassette_name: 'data_gouv_sirene_july_KO' } do
+  context 'when HTTP interaction fails', vcr: { cassette_name: 'cquest_geo_sirene_may_KO' } do
     before do
       allow_any_instance_of(described_class)
         .to receive(:base_uri)
